fix: upgrade to v16 to fix security issue. It required upgrade of next, react and material-ui. Next middleware is moved to proxy.ts and legacy image component is removed.
