+++ frontend/next.config.ts
@@ -1,14 +1,14 @@
 // SPDX-FileCopyrightText: 2021 - 2023 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2021 - 2023 dv4all
 // SPDX-FileCopyrightText: 2022 - 2025 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
-// SPDX-FileCopyrightText: 2022 - 2025 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2022 - 2026 Netherlands eScience Center
 // SPDX-FileCopyrightText: 2022 Jesús García Gonzalez (Netherlands eScience Center) <j.g.gonzalez@esciencecenter.nl>
-// SPDX-FileCopyrightText: 2023 - 2025 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 - 2026 Dusan Mijatovic (Netherlands eScience Center)
 //
 // SPDX-License-Identifier: Apache-2.0
 
 /** @type {import('next').NextConfig} */
-import type {NextConfig} from 'next'
+import type {NextConfig} from 'next/types'
 
 import rewritesConfig from './next.rewrites'
 import securityHeaders from './next.headers'
@@ -25,11 +25,12 @@ const nextConfig: NextConfig = {
   // disable strict mode if you want react to render component once
   // see for more info https://nextjs.org/docs/api-reference/next.config.js/react-strict-mode
   reactStrictMode: false,
-  eslint: {
-    // Run ESLint in these directories during production builds (next build)
-    // by default next runs linter only in pages/, components/, and lib/
-    dirs: ['app','auth', 'components', 'config', 'pages', 'styles', 'types', 'utils']
-  },
+  // functionality removed in v16
+  // eslint: {
+  //   // Run ESLint in these directories during production builds (next build)
+  //   // by default next runs linter only in pages/, components/, and lib/
+  //   dirs: ['app','auth', 'components', 'config', 'pages', 'styles', 'types', 'utils']
+  // },
   // only in development
   rewrites: async () => rewritesConfig,
