+++ frontend/package-lock.json.license
@@ -1,5 +1,5 @@
-SPDX-FileCopyrightText: 2024 - 2025 Dusan Mijatovic (Netherlands eScience Center)
-SPDX-FileCopyrightText: 2024 - 2025 Netherlands eScience Center
+SPDX-FileCopyrightText: 2024 - 2026 Dusan Mijatovic (Netherlands eScience Center)
+SPDX-FileCopyrightText: 2024 - 2026 Netherlands eScience Center
 SPDX-FileCopyrightText: 2024 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
 SPDX-FileCopyrightText: 2025 Dusan Mijatovic (dv4all) (dv4all)
 SPDX-FileCopyrightText: 2025 dv4all
