+++ frontend/package.json
@@ -32,18 +32,18 @@
     "@emotion/react": "11.14.0",
     "@emotion/server": "11.11.0",
     "@emotion/styled": "11.14.1",
-    "@mui/icons-material": "7.3.4",
-    "@mui/material": "7.3.4",
-    "@mui/material-nextjs": "7.3.3",
+    "@mui/icons-material": "7.3.6",
+    "@mui/material": "7.3.6",
+    "@mui/material-nextjs": "7.3.6",
     "@tailwindcss/typography": "0.5.19",
     "aos": "2.3.4",
     "cookie": "1.0.2",
     "d3": "7.9.0",
     "dompurify": "3.2.7",
     "jsonwebtoken": "9.0.2",
-    "next": "15.5.7",
-    "react": "19.2.1",
-    "react-dom": "19.2.1",
+    "next": "16.1.1",
+    "react": "19.2.3",
+    "react-dom": "19.2.3",
     "react-hook-form": "7.63.0",
     "react-markdown": "10.1.0",
     "rehype-external-links": "3.0.0",
@@ -65,8 +65,8 @@
     "@types/node": "24.6.2",
     "@types/react": "19.2.7",
     "@types/react-dom": "19.2.3",
-    "eslint": "9.37.0",
-    "eslint-config-next": "15.5.4",
+    "eslint": "9.39.2",
+    "eslint-config-next": "16.1.1",
     "jest": "30.2.0",
     "jest-environment-jsdom": "30.2.0",
     "knip": "5.64.1",
