+++ frontend/tsconfig.json.license
@@ -1,8 +1,8 @@
 SPDX-FileCopyrightText: 2021 - 2022 Dusan Mijatovic (dv4all)
 SPDX-FileCopyrightText: 2021 - 2022 dv4all
-SPDX-FileCopyrightText: 2022 - 2025 Netherlands eScience Center
+SPDX-FileCopyrightText: 2022 - 2026 Netherlands eScience Center
 SPDX-FileCopyrightText: 2022 Jesús García Gonzalez (Netherlands eScience Center) <j.g.gonzalez@esciencecenter.nl>
-SPDX-FileCopyrightText: 2024 - 2025 Dusan Mijatovic (Netherlands eScience Center)
+SPDX-FileCopyrightText: 2024 - 2026 Dusan Mijatovic (Netherlands eScience Center)
 
 SPDX-License-Identifier: Apache-2.0
 SPDX-License-Identifier: CC-BY-4.0
